feat: add React Router support and deprecate `remix` function in favor of `reactRouter`
